- 27.02.2024 güncellemeler
- Merge remote-tracking branch 'origin/master' into HuseyinBranch
- Üretim ekranı ana sayfa duruş entegrasyonu tamamlandı.
- merge c
- 14.03.2024 Üretim ekran yazılımı devam ediyor.

+++ TsiErp.ErpUI/Pages/MaintenanceManagement/MaintenanceMaterialPurchaseTracking/MaintenanceMaterialPurchaseTrackingsListPage.razor.cs
@@ -25,8 +25,8 @@ namespace TsiErp.ErpUI.Pages.MaintenanceManagement.MaintenanceMaterialPurchaseTr
             public string PurchaseCurrentAccountName { get; set; }
             public string ProductCode { get; set; }
             public string ProductName { get; set; }
-            public int RequirementAmount { get; set; }
-            public int OrderAmount { get; set; }
+            public decimal RequirementAmount { get; set; }
+            public decimal OrderAmount { get; set; }
             public string SupplyStatus { get; set; }
             public DateTime? SupplyDate { get; set; }
